Fix the interpretor unit test and refine the test file name
