+++ query/src/interpreters/interpreter_factory.rs
@@ -19,8 +19,6 @@ use common_exception::Result;
 use common_planners::PlanNode;
 
 use super::DescribeUserStageInterpreter;
-use crate::interpreters::interpreter_table_optimize::OptimizeTableInterpreter;
-use crate::interpreters::interpreter_user_stage_drop::DropUserStageInterpreter;
 use crate::interpreters::AlterUserInterpreter;
 use crate::interpreters::AlterUserUDFInterpreter;
 use crate::interpreters::CopyInterpreter;
@@ -33,6 +31,7 @@ use crate::interpreters::DescribeTableInterpreter;
 use crate::interpreters::DropDatabaseInterpreter;
 use crate::interpreters::DropTableInterpreter;
 use crate::interpreters::DropUserInterpreter;
+use crate::interpreters::DropUserStageInterpreter;
 use crate::interpreters::DropUserUDFInterpreter;
 use crate::interpreters::ExplainInterpreter;
 use crate::interpreters::GrantPrivilegeInterpreter;
@@ -40,6 +39,7 @@ use crate::interpreters::InsertInterpreter;
 use crate::interpreters::InterceptorInterpreter;
 use crate::interpreters::Interpreter;
 use crate::interpreters::KillInterpreter;
+use crate::interpreters::OptimizeTableInterpreter;
 use crate::interpreters::RevokePrivilegeInterpreter;
 use crate::interpreters::SelectInterpreter;
 use crate::interpreters::SettingInterpreter;
