+++ query/tests/it/interpreters/interpreter_admin_use_tenant.rs
