+++ query/tests/it/interpreters/interpreter_table_show_create.rs
