+++ query/tests/it/interpreters/interpreter_user_previlege_revoke.rs
