+++ query/tests/it/interpreters/interpreter_user_stage_create.rs
@@ -33,7 +33,7 @@ async fn test_create_stage_interpreter() -> Result<()> {
     {
         let plan = PlanParser::parse(ctx.clone(), query).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "CreatStageInterpreter");
+        assert_eq!(executor.name(), "CreateUserStageInterpreter");
         executor.execute(None).await?;
         let stage = ctx
             .get_user_manager()
@@ -52,7 +52,7 @@ async fn test_create_stage_interpreter() -> Result<()> {
     {
         let plan = PlanParser::parse(ctx.clone(), query).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "CreatStageInterpreter");
+        assert_eq!(executor.name(), "CreateUserStageInterpreter");
         executor.execute(None).await?;
 
         let stage = ctx
@@ -72,7 +72,7 @@ async fn test_create_stage_interpreter() -> Result<()> {
         let query = "CREATE STAGE test_stage url='s3://load/files/' credentials=(access_key_id='1a2b3c' secret_access_key='4x5y6z') file_format=(FORMAT=CSV compression=GZIP record_delimiter='\n') comments='test'";
         let plan = PlanParser::parse(ctx.clone(), query).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "CreatStageInterpreter");
+        assert_eq!(executor.name(), "CreateUserStageInterpreter");
         let r = executor.execute(None).await;
         assert!(r.is_err());
         let e = r.err();
