+++ query/tests/it/interpreters/interpreter_user_udf_alter.rs
@@ -31,7 +31,7 @@ async fn test_alter_udf_interpreter() -> Result<()> {
             "CREATE FUNCTION IF NOT EXISTS isnotempty AS (p) -> not(isnull(p)) DESC = 'This is a description'";
         let plan = PlanParser::parse(ctx.clone(), query).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "CreatUDFInterpreter");
+        assert_eq!(executor.name(), "CreateUserUDFInterpreter");
         let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
         let udf = ctx
@@ -50,7 +50,7 @@ async fn test_alter_udf_interpreter() -> Result<()> {
         "ALTER FUNCTION isnotempty AS (d) -> not(isnotnull(d)) DESC = 'This is a new description'";
         let plan = PlanParser::parse(ctx.clone(), query).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "AlterUDFInterpreter");
+        assert_eq!(executor.name(), "AlterUserUDFInterpreter");
 
         let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
