+++ query/tests/it/interpreters/interpreter_user_udf_drop.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 use common_base::tokio;
-use common_exception::ErrorCode;
 use common_exception::Result;
 use databend_query::interpreters::*;
 use databend_query::sql::*;
@@ -21,19 +20,21 @@ use futures::stream::StreamExt;
 use pretty_assertions::assert_eq;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_create_udf_interpreter() -> Result<()> {
+async fn test_drop_udf_interpreter() -> Result<()> {
     common_tracing::init_default_ut_tracing();
 
     let ctx = crate::tests::create_query_context()?;
     let tenant = ctx.get_tenant();
 
-    let query =
+    static CREATE_UDF: &str =
         "CREATE FUNCTION IF NOT EXISTS isnotempty AS (p) -> not(isnull(p)) DESC = 'This is a description'";
 
+    static DROP_UDF_IF_EXISTS: &str = "DROP FUNCTION IF EXISTS isnotempty";
+    static DROP_UDF: &str = "DROP FUNCTION isnotempty";
+
     {
-        let plan = PlanParser::parse(ctx.clone(), query).await?;
+        let plan = PlanParser::parse(ctx.clone(), CREATE_UDF).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        assert_eq!(executor.name(), "CreatUDFInterpreter");
         let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
         let udf = ctx
@@ -48,32 +49,35 @@ async fn test_create_udf_interpreter() -> Result<()> {
     }
 
     {
-        // IF NOT EXISTS.
-        let plan = PlanParser::parse(ctx.clone(), query).await?;
+        let plan = PlanParser::parse(ctx.clone(), DROP_UDF).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        executor.execute(None).await?;
-
-        let udf = ctx
-            .get_user_manager()
-            .get_udf(&tenant, "isnotempty")
-            .await?;
-
-        assert_eq!(udf.name, "isnotempty");
-        assert_eq!(udf.parameters, vec!["p".to_string()]);
-        assert_eq!(udf.definition, "not(isnull(p))");
-        assert_eq!(udf.description, "This is a description")
+        assert_eq!(executor.name(), "DropUserUDFInterpreter");
+        let res = executor.execute(None).await;
+        assert!(res.is_ok());
     }
 
     {
-        let query1 =
-            "CREATE FUNCTION isnotempty AS (p) -> not(isnull(p)) DESC = 'This is a description'";
-        let plan = PlanParser::parse(ctx.clone(), query1).await?;
+        let plan = PlanParser::parse(ctx.clone(), DROP_UDF_IF_EXISTS).await?;
         let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
-        let r = executor.execute(None).await;
-        assert!(r.is_err());
-        let e = r.err();
-        assert_eq!(e.unwrap().code(), ErrorCode::udf_already_exists_code());
+        assert_eq!(executor.name(), "DropUserUDFInterpreter");
+        let res = executor.execute(None).await;
+        assert!(res.is_ok());
+    }
 
+    {
+        let plan = PlanParser::parse(ctx.clone(), DROP_UDF).await?;
+        let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
+        assert_eq!(executor.name(), "DropUserUDFInterpreter");
+        let res = executor.execute(None).await;
+        assert!(res.is_err());
+    }
+
+    {
+        let plan = PlanParser::parse(ctx.clone(), CREATE_UDF).await?;
+        let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
+        assert_eq!(executor.name(), "CreateUserUDFInterpreter");
+        let mut stream = executor.execute(None).await?;
+        while let Some(_block) = stream.next().await {}
         let udf = ctx
             .get_user_manager()
             .get_udf(&tenant, "isnotempty")
@@ -84,5 +88,13 @@ async fn test_create_udf_interpreter() -> Result<()> {
         assert_eq!(udf.definition, "not(isnull(p))");
         assert_eq!(udf.description, "This is a description")
     }
+    {
+        let plan = PlanParser::parse(ctx.clone(), DROP_UDF_IF_EXISTS).await?;
+        let executor = InterpreterFactory::get(ctx.clone(), plan.clone())?;
+        assert_eq!(executor.name(), "DropUserUDFInterpreter");
+        let res = executor.execute(None).await;
+        assert!(res.is_ok());
+    }
+
     Ok(())
 }
