+++ query/tests/it/interpreters/mod.rs
@@ -13,30 +13,30 @@
 // limitations under the License.
 
 mod access;
+mod interpreter_admin_use_tenant;
 mod interpreter_database_create;
 mod interpreter_database_drop;
-mod interpreter_describe_stage;
-mod interpreter_describe_table;
+mod interpreter_database_show_create;
 mod interpreter_explain;
-mod interpreter_grant_privilege;
+mod interpreter_factory_interceptor;
 mod interpreter_insert;
-mod interpreter_interceptor;
-mod interpreter_revoke_previlege;
 mod interpreter_select;
 mod interpreter_setting;
-mod interpreter_show_create_database;
-mod interpreter_show_create_table;
-mod interpreter_stage_create;
-mod interpreter_stage_drop;
 mod interpreter_table_create;
+mod interpreter_table_describe;
 mod interpreter_table_drop;
+mod interpreter_table_show_create;
 mod interpreter_table_truncate;
-mod interpreter_udf_alter;
-mod interpreter_udf_create;
-mod interpreter_udf_drop;
 mod interpreter_use_database;
-mod interpreter_use_tenant;
 mod interpreter_user_alter;
 mod interpreter_user_create;
 mod interpreter_user_drop;
+mod interpreter_user_previlege_revoke;
+mod interpreter_user_privilege_grant;
+mod interpreter_user_stage_create;
+mod interpreter_user_stage_describe;
+mod interpreter_user_stage_drop;
+mod interpreter_user_udf_alter;
+mod interpreter_user_udf_create;
+mod interpreter_user_udf_drop;
 mod plan_schedulers;
